% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet-package.R
\docType{package}
\name{HomUHet-package}
\alias{HomUHet-package}
\alias{_PACKAGE}
\alias{{HomUHet}-package}
\alias{HomUHet}
\alias{can}
\alias{be}
\alias{used}
\alias{to}
\alias{identify}
\alias{and}
\alias{separates}
\alias{predictors}
\alias{with}
\alias{homogeneous}
\alias{or}
\alias{heterogeneous}
\alias{effect}
\alias{across}
\alias{multiple}
\alias{data}
\alias{sets}
\alias{through}
\alias{fitting}
\alias{penalized}
\alias{linear}
\alias{regression}
\alias{models}
\alias{in}
\alias{two}
\alias{steps.}
\alias{Applicable}
\alias{Gaussian}
\alias{response}
\alias{variable}
\alias{very}
\alias{high}
\alias{dimensional}
\alias{where}
\alias{the}
\alias{number}
\alias{of}
\alias{could}
\alias{larger}
\alias{than}
\alias{observations}
\alias{per}
\alias{set.}
\title{identify homogeneous and heterogeneous effects of predictors in multiple datasets}
\description{
classify the predictors into homogeneous, heterogeneous and unassociated categories. outputs the solution path plots.
}
\author{
\strong{Maintainer}: Pei S. Yang \email{yang.1736@osu.edu}

}
\keyword{internal}
