% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\name{HomUHet.plot}
\alias{HomUHet.plot}
\title{Plot the solution path from HomUHet}
\usage{
HomUHet.plot(fit, y_name = NULL)
}
\arguments{
\item{fit}{the output object from HomUHet}

\item{y_name}{if needed, a response variable name for the solution path plots. Default is NULL.}
}
\value{
solution path plots from Step 1 and Step 2
}
\description{
This function outputs the solution path plots
}
