% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\name{HomUHet}
\alias{HomUHet}
\title{HomUHet}
\usage{
HomUHet(data, sid, solution_path = FALSE, y_name = NULL)
}
\arguments{
\item{data}{the data containing observations from all studies, with the response variable in the first column, followed by the predictor matrix.}

\item{sid}{the study id for each observation in data}

\item{solution_path}{TRUE if the user wishes to output the solution path plots. Default is FALSE}

\item{y_name}{if needed, a response variable name for the solution path plots. Default is NULL.}
}
\value{
names of Homogeneous and Heterogeneous predictors, estimates of predictors, solution path plots
}
\description{
This function outputs the names of predictors with homogeneous or heterogeneous predictors across multiple data sets, the estimates of predictors, and solution plots
}
