% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\name{HomUHet.sim.beta}
\alias{HomUHet.sim.beta}
\title{simulates homogeneous and heterogeneous coefficients of predictors}
\usage{
HomUHet.sim.beta(J, K, n_homo, n_heter, level = c("l", "m", "h"))
}
\arguments{
\item{J}{The total number of predictors including the predictors with homogeneous and heterogeneous effects and the predictors without effects}

\item{K}{The number of studies}

\item{n_homo}{the number of homogeneous coefficients}

\item{n_heter}{the number of heterogeneous coefficients}

\item{level}{the level of heterogeneity in the heterogeneous coefficients,
where "l" stands for low, "m" stands for medium and "h" stands for high.}
}
\value{
the simulated coefficient matrix and miscellenance information about it
\item{beta}{the K x J coefficient matrix}
\item{J}{the number of predictors including both predictors which have effects and which do not}
\item{K}{the number of studies}
\item{homo_index}{a vector containing the column numbers of homogeneous coefficients in the coefficient matrix}
\item{heter_index}{a vector containing the column numbers of homogeneous coefficients in the coefficient matrix}
}
\description{
this function outputs matrix of coefficients of predictors
}
