% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet_data.R
\docType{data}
\name{HomUHet_data}
\alias{HomUHet_data}
\title{simulated data to be used as demonstration for HomUHet package}
\format{
A data frame with 444 rows and 502 variables:
\describe{
\item{Study_label}{integers indicating the study number}
\item{y}{the response variable}
...
}
}
\usage{
HomUHet_data
}
\description{
A dataset containing observations of 4 studies and 500 predictors
}
\keyword{datasets}
