% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\docType{package}
\name{HomUHet-package}
\alias{HomUHet}
\alias{HomUHet-package}
\title{identify homogeneous and heterogeneous effects of predictors in multiple datasets}
\description{
classify the predictors into homogeneous, heterogeneous and unassociated categories. outputs the solution path plots.
}
\details{
The only function you're likely to need from \pkg{HomUHet} is
\code{\link{HomUHet}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\author{
\strong{Maintainer}: Pei S. Yang \email{yang.1736@osu.edu}

}
