% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\docType{package}
\name{HomUHet-package}
\alias{HomUHet}
\alias{HomUHet-package}
\title{identify homogeneous and heterogeneous effects of predictors in multiple datasets}
\description{
classify the predictors into homogeneous, heterogeneous and unassociated categories. outputs the solution path plots.
}
\details{
HomUHet can be used to identify and separates predictors with homogeneous or heterogeneous effect across multiple data sets
through fitting penalized linear regression models in two steps.
Applicable to Gaussian response variable and very high dimensional data
where the number of predictors could be larger than the number of observations per data set.
}
\author{
\strong{Maintainer}: Pei S. Yang \email{yang.1736@osu.edu}

}
