% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\name{HomUHet}
\alias{HomUHet}
\title{fit a two-step penalized regression model}
\usage{
HomUHet(x, y, sid, solution_path_plot = FALSE)
}
\arguments{
\item{x}{the predictor matrix. a matrix of n x J containing observations from all studies for all predictors}

\item{y}{the response variable following gaussian distribution. a vector of n observations for the response variable}

\item{sid}{a vector of integers indexing the study id for each observation in data}

\item{solution_path_plot}{TRUE if outputting solution path plots is desired}
}
\value{
the names of identified predictors and their estimated effects
\item{Homo}{a character string of names of homogeneous predictors}
\item{Heter}{a character string of N=names of heterogeneous predictors}
\item{coefficients}{estimated coefficients of the homogeneous and heterogeneous predictors in K studies}
}
\description{
This function outputs the names of predictors with homogeneous or heterogeneous predictors across multiple data sets, the estimates of predictors, and solution plots
}
