% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomUHet.R
\name{HomUHet}
\alias{HomUHet}
\title{fit a two-step penalized regression model}
\usage{
HomUHet(data, solution_path_plot = FALSE)
}
\arguments{
\item{data}{The input dataframe containing observations from all studies where the first column is the study label,
the second column is the response variable and the following columns are the predictors.}

\item{solution_path_plot}{TRUE if outputting solution path plots is desired}
}
\value{
the names of identified predictors and their estimated effects
\item{Homo}{a character string of names of homogeneous predictors}
\item{Heter}{a character string of N=names of heterogeneous predictors}
\item{coefficients}{a data frame containing estimated coefficients of the homogeneous and heterogeneous predictors in K studies}
}
\description{
This function outputs the names of predictors with homogeneous or heterogeneous predictors across multiple data sets, the estimates of predictors, and solution plots
}
